requirements.txt
# Pinned on Python 3.10 — 2026-04-11
# Note: GPU/large packages (torch, transformers, gradio, faiss-cpu, etc.)
# are left unpinned — pin after local GPU environment install.

# Pinned
fastapi==0.135.3
uvicorn==0.44.0
pydantic==2.12.5
python-json-logger==4.1.0
mlflow==3.11.1
pyyaml==6.0.3
numpy==2.2.6
pandas==2.3.3
pyarrow==23.0.1
scipy==1.15.3
httpx==0.28.1
psutil==7.2.2
safetensors==0.7.0
scikit-learn==1.7.2
pandera==0.30.1
slowapi==0.1.9
prometheus-fastapi-instrumentator==7.1.0
sse-starlette==3.3.4

# Unpinned (GPU/large — pin after local install)
torch
transformers
peft
datasets
accelerate
gradio>=5.0
plotly>=5.22.0
faiss-cpu
sentence-transformers
captum
confusable-homoglyphs
onnxruntime
optimum
